# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fortexa Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
  /deposits:
    get:
      operationId: getDeposits
      tags: [deposits]
      summary: List user deposits
      responses:
        "200":
          description: List of deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Create a deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Withdrawals ─────────────────────────────────────────────────────────────
  /withdrawals:
    get:
      operationId: getWithdrawals
      tags: [withdrawals]
      summary: List user withdrawals
      responses:
        "200":
          description: List of withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Create a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Transactions ────────────────────────────────────────────────────────────
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /profile/password:
    put:
      operationId: updatePassword
      tags: [profile]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordUpdate"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Announcements ───────────────────────────────────────────────────────────
...
  /admin/withdrawals:
    get:
      operationId: getAdminWithdrawals
      tags: [admin]
      summary: List all withdrawal requests
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: limit
          schema:
            type: number
            default: 50
        - in: query
          name: offset
          schema:
            type: number
            default: 0
      responses:
        "200":
          description: Withdrawals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTransactionList"
  /admin/withdrawals/{id}:
    put:
      operationId: updateAdminWithdrawal
      tags: [admin]
      summary: Approve or reject a withdrawal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ── Auth ──
    RegisterInput:
      type: object
      required: [name, phone, email, password]
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
        email:
          type: string
          
        password:
          type: string
          minLength: 6
        referralCode:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, phone, email, referralCode, investmentBalance, gainBalance, status, role, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        referralCode:
          type: string
        referredById:
          type: ["number", "null"]
        investmentBalance:
          type: number
        gainBalance:
          type: number
        status:
          type: string
          enum: [active, suspended, banned]
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time

    # ── Dashboard ──
...
    Transaction:
      type: object
      required: [id, userId, type, amount, fee, netAmount, status, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        type:
          type: string
          enum: [deposit, withdrawal, commission, gain]
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        description:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DepositInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 3000
    WithdrawalInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 3000

    # ── Referrals ──
...
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
        email:
          type: string
    PasswordUpdate:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6

    # ── Announcements ──
...
    AdminTransactionList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminTransaction"
        total:
          type: number
    AdminTransaction:
      type: object
      required: [id, userId, userName, userEmail, type, amount, fee, netAmount, status, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        userName:
          type: string
        userEmail:
          type: string
        type:
          type: string
          enum: [deposit, withdrawal, commission, gain]
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        description:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TransactionStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        amount:
          type: ["number", "null"]
        rejectionReason:
          type: ["string", "null"]
...